Use "useSignup" hook
